+++ components/lib/organizationchart/OrganizationChartBase.js
@@ -71,7 +71,7 @@ const classes = {
     nodes: 'p-organizationchart-nodes',
     lines: 'p-organizationchart-lines',
     lineLeft: ({ index }) => classNames('p-organizationchart-line-left', { 'p-organizationchart-line-top': index !== 0 }),
-    lineRight: ({ index }) => classNames('p-organizationchart-line-right', { 'p-organizationchart-line-top': index !== nodeChildLength - 1 }),
+    lineRight: ({ index, nodeChildLength }) => classNames('p-organizationchart-line-right', { 'p-organizationchart-line-top': index !== nodeChildLength - 1 }),
     lineDown: 'p-organizationchart-line-down',
     nodeTogglerIcon: 'p-node-toggler-icon',
     nodeToggler: 'p-node-toggler'
